add compress, refactor some codes
